backend/requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
pydantic
python-multipart
python-dotenv
